Share SSH agent option parsing
